+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 HEAD
 ----
 
+* Warn on missing ``FILTER`` entries on parsing
 * Reordered parameters in ``from_stream`` and ``from_file`` (#18)
 * Renamed ``from_file`` to ``from_stream`` (#18)
 * Renamed ``Reader.jump_to`` to ``Reader.fetch``